fix: int to size_t

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: jungao <jungao@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/16 23:29:56 by jungao            #+#    #+#             */
-/*   Updated: 2021/05/17 00:44:20 by jungao           ###   ########.fr       */
+/*   Updated: 2021/05/20 20:01:12 by jungao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*dst;
-	int				i;
+	size_t			i;
 
 	dst = (unsigned char *)s;
 	i = 0;
